+++ src/geolocation/Map.js
@@ -3,26 +3,28 @@ import PropTypes from 'prop-types'
 import debounce from 'lodash/debounce'
 import geolocationAutoCompleteAddress from './geolocationAutoCompleteAddress'
 import { injectRules } from '../addressRulesContext'
+import { compose } from 'recompose'
+import { injectAddressContext } from '../addressContainerContext'
+import AddressShapeWithValidation from '../propTypes/AddressShapeWithValidation'
 
 class Map extends Component {
-  constructor(props) {
-    super(props)
-
-    this.mapMounted = this.mapMounted.bind(this)
+  getCoordinatesFromProps(props) {
+    const { geoCoordinates, address } = props
+    return geoCoordinates || address.geoCoordinates.value
   }
 
   shouldComponentUpdate(prevProps) {
     const rulesChanged = prevProps.rules.country !== this.props.rules.country
     const geoCoordsChanged = this.isDifferentGeoCoords(
-      prevProps.geoCoordinates,
-      this.props.geoCoordinates,
+      this.getCoordinatesFromProps(prevProps),
+      this.getCoordinatesFromProps(this.props),
     )
 
     return geoCoordsChanged || rulesChanged
   }
 
   componentDidUpdate() {
-    const location = this.getLocation(this.props.geoCoordinates)
+    const location = this.getLocation()
     this.changeMarkerPosition(location)
     this.recenterMap(location)
   }
@@ -35,7 +37,7 @@ class Map extends Component {
       return
     }
 
-    const location = this.getLocation(this.props.geoCoordinates)
+    const location = this.getLocation()
     this.createMap(mapElement, location)
     this.changeMarkerPosition(location)
   }
@@ -83,8 +85,8 @@ class Map extends Component {
     )
   }
 
-  getLocation = geoCoordinates => {
-    const [lng, lat] = this.props.geoCoordinates
+  getLocation = () => {
+    const [lng, lat] = this.getCoordinatesFromProps(this.props)
     const location = new this.props.googleMaps.LatLng(lat, lng)
     return location
   }
@@ -148,11 +150,16 @@ Map.defaultProps = {
 Map.propTypes = {
   loadingElement: PropTypes.node,
   mapProps: PropTypes.object,
-  geoCoordinates: PropTypes.array.isRequired,
+  geoCoordinates: PropTypes.array,
+  address: AddressShapeWithValidation,
   rules: PropTypes.object.isRequired,
   onChangeAddress: PropTypes.func.isRequired,
   loadingGoogle: PropTypes.bool,
   googleMaps: PropTypes.object,
 }
 
-export default injectRules(Map)
+const enhance = compose(
+  injectAddressContext,
+  injectRules,
+)
+export default enhance(Map)