+++ CHANGELOG.md
@@ -7,11 +7,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.0.1] - 2018-07-26
+
+## Changed
+
+- Updated README to include `v2.0.0` changes
+
+## [2.0.0] - 2018-07-26
+
 ### Changed
 
 - **[BREAKING]** Both inputs moved to an `inputs` folder
 - **[BREAKING]** CustomInput renamed to StyleguideInput [[#103](https://github.com/vtex/address-form/pull/103)]
+- **[BREAKING]** AddressContainer now uses plain children instead of render prop
+- AddressContainer now passes down `address`, `onChangeAddress` and `Input` via context to children (can be overriden in each component for backwards compatibility)
+- Map now does not require the `geoCoordinates` prop; it may use the coordinates from the injected address as default [[#104](https://github.com/vtex/address-form/pull/104)]
 
 ### Added
 
 - This Changelog
+- **`AddressSubmitter`** component and tests